% Auto-generated: do not edit by hand
\name{''DashStickyContainer}

\alias{''DashStickyContainer}

\title{DashStickyContainer component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashStickyContainer(children=NULL, id=NULL, className=NULL, title=NULL, top=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{title}{Character. }

\item{top}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

